Set up lefthook + ruff and apply project-wide pass
Dev tooling:
- ruff + lefthook added to the pixi dev feature
- pyproject.toml gains [tool.ruff] (line-length 100, py312, select
  E/W/F/I/B/UP/SIM/RUF, ignore B008 for typer Argument/Option defaults,
  ignore RUF001/RUF002 for intentional × in UI labels) and matching
  format config
- lefthook.yml runs `ruff check --fix` then `ruff format` on staged
  *.py files at pre-commit, with stage_fixed so autofixes land in the
  same commit
- New pixi tasks: lint, format, install-hooks
- README documents the one-time `install-hooks` step

Codebase pass:
- Import re-sorting (isort via ruff) across the tree
- StrEnum for JobState instead of (str, Enum) — UP042
- Drop redundant int() around round() — RUF046
- Combine nested if into a single conjunction in VideoCanvas — SIM102
- Combine nested with blocks in test_landing — SIM117
- Replace Optional[X] / typing.Iterable with X | None /
  collections.abc.Iterable (UP)
- Remove unused pytest / info imports
- ruff format normalizes whitespace; no behavior changes

104/104 tests still pass after the rewrite.

+++ src/croppy/ffmpeg/probe.py
@@ -75,9 +75,7 @@ def probe(path: Path | str) -> VideoInfo:
     height = int(stream["height"])
     codec = str(stream.get("codec_name", ""))
     container = str(fmt.get("format_name", ""))
-    fps = _parse_rate(
-        stream.get("avg_frame_rate") or stream.get("r_frame_rate") or "0/0"
-    )
+    fps = _parse_rate(stream.get("avg_frame_rate") or stream.get("r_frame_rate") or "0/0")
     duration = _parse_duration(stream, fmt)
     nb_frames = _parse_nb_frames(stream, duration, fps)
 
@@ -113,9 +111,7 @@ def _parse_duration(stream: dict[str, Any], fmt: dict[str, Any]) -> float:
     return 0.0
 
 
-def _parse_nb_frames(
-    stream: dict[str, Any], duration: float, fps: float
-) -> int | None:
+def _parse_nb_frames(stream: dict[str, Any], duration: float, fps: float) -> int | None:
     val = stream.get("nb_frames")
     if val is not None:
         try:
@@ -123,5 +119,5 @@ def _parse_nb_frames(
         except (TypeError, ValueError):
             pass
     if duration > 0 and fps > 0:
-        return int(round(duration * fps))
+        return round(duration * fps)
     return None
